Revert "feat: updated behaviors"
This reverts commit cf64a1a66704e089253216aec69cf3da9bd2a291.
