+++ examples/crud-app/Crud.Core/UseCases/Create.cs
@@ -13,9 +13,10 @@ public sealed class CreateQuestionDependencies : IFeatureDependencies
     {
         featureBuilder
             .AddEndpoint<EndpointDefinition>()
-            .AddLoggingFor<Command>().UsingSpanishTemplate()
-            .AddFluentValidationUsing<Validator>()
-            .AddExceptionHandling<LoggingExceptionHandlerPipeline<Command, Unit>>()
+            .AddLoggingBehaviorFor<Command>()
+                .UsingSpanishTemplate()
+            .AddFluentValidationBehaviorUsing<Validator>()
+            .AddExceptionHandlingBehavior<LoggingExceptionHandlerPipeline<Command, Unit>>()
             .AddHandler<Handler>();
     }
 }
