+++ examples/crud-app/Crud.CrossCutting/Crud.CrossCutting.csproj
@@ -20,7 +20,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\src-app\VSlices.CrossCutting.BackgroundTaskListener.Hosting\VSlices.CrossCutting.BackgroundTaskListener.Hosting.csproj" />
-    <ProjectReference Include="..\..\..\src-app\VSlices.CrossCutting.Pipeline.EventFiltering\VSlices.CrossCutting.Pipeline.EventFiltering.csproj" />
     <ProjectReference Include="..\..\..\src-app\VSlices.CrossCutting.Pipeline.ExceptionHandling\VSlices.CrossCutting.Pipeline.ExceptionHandling.csproj" />
     <ProjectReference Include="..\..\..\src-app\VSlices.CrossCutting.Pipeline.FluentValidation\VSlices.CrossCutting.Pipeline.FluentValidation.csproj" />
     <ProjectReference Include="..\..\..\src-app\VSlices.CrossCutting.Pipeline.Logging\VSlices.CrossCutting.Pipeline.Logging.csproj" />
